go/analysis/passes/printf: use induction on interface methods
This CL causes the printf analyzer to use the presence of
an assignment from an implementation to an interface type

 type I interface { Printf(...) }
 type T ...
 var _ I = T(...)

as evidence that: if a method T.F is found to be a
printf wrapper, then the corresponding abstract method
I.F is also to be considered a printf wrapper.
This allows induction through interfaces without the
need for annotations.

Only assigments within the same package as I.Printf are
considered.

This feature is only enabled for interface methods declared
in files using at least go1.26, to avoid surprises due to
go test's vet suite getting stricter.

+ docs, tests

Fixes golang/go#76368
For golang/go#58340

Change-Id: I124aa383d9b4278ffe8d982ed630813a836bd0cf
Reviewed-on: https://go-review.googlesource.com/c/tools/+/721984
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>