internal/refactor/inline: simplify import handling
Previously, the inliner had a very heavyweight API that
accepted a whole file []byte as input and produced a whole
fixed file []byte as output, and used various whole-file
parse and format steps along the way. (Also, the clients
still had to compute the diff afterwards!)

Not only was this expensive, showing up in profiles
(see golang/go#75773), but it was unnecessary: by construction,
the only edits due to inlining are at the call site (or
some ancestor) and around the import declaration. The new
logic reflects this. The new API accepts only the AST, and
returns only a list of []refactor.Edit.

(refactor.Edit is an alias for analysis.TextEdit. Ideally
it would be a distinct isomorphic type without a dependency
between the two packages, but slices of different named
types are cumbersome to convert between.)

The import computation is now much simpler:
- Addition uses refactor.AddImportEdits (the bottom half
  of refactor.AddImport), which does the surgery without the
  analytical validation.
- Import deletion is done by new ad-hoc logic, but only
  for the packages made redundant from call position (pkg.F(...)),
  since it is common to replace dir1/pkg by dir2/pkg and
  we don't want to leave an ambiguous import. For all other
  packages that are made redundant by the transformation
  (e.g. mentioned in parameter type decls), we let the
  client remove them, which is something that the analysis
  drivers and tests now do, freeing the inliner from having
  to compute unused imports generally.
  (gopls' inline-all code action also does it after each step.)

Also:
- The inliner now rejects all callers in generated files.
  (Previously it was only from cgo files, which required
  the whole file content to detect.)
- The inconsistent offsets assertion is gone.
  (It required the whole file content.)
- Added test case import-comments2.txtar (from golang/go#67336)
  to inline/, not just gopls. The new behavior actually fixes
  comment migration, which was oddly not tested by the CL that
  fixed that issue, given that it was in the title of the bug.
- Fixed a bug in AddImport (re: 'before') that caused it to
  create a new import decl if the existing one had a comment.
- Fixed some @codeaction tests now that we don't gofmt or
  goimports (=> no import decl merging)
- Tweaked got/want/diff section headers in test frameworks
  for ease of reading.

Updates golang/go#67336
Fixes golang/go#76358

Change-Id: I223747aed58b3287df2494df4814b7982ccc9f77
Reviewed-on: https://go-review.googlesource.com/c/tools/+/723781
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>