gopls/internal/cache: make unimported completions deterministic
Completions found in the module cache are accumulated in a map,
so that, without extra processing, the result can be nondeterministic.
This CL fixes that, and also ensures that neither vendored nor
internal packages will be recommended.

This was tested by using elaborated versions of the example from
the bug report.

Completes the fix of golang/go#76310

Change-Id: I6bbc1f5737c4c5b70d1b6f50b250f76becb7ef95
Reviewed-on: https://go-review.googlesource.com/c/tools/+/723540
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>