go/analysis/passes/modernize: rangeint: handle usages of loop label
Currently, when a label is used, the rangeint modernizer fails to
properly detect when i is used after a loop. This is because we only
search for i usage within the for loop's parent subtree. When a label is
used, the label becomes the loop's parent, and i can only be found in
the loop's loop's parent subtree.

To fix this, this change makes the rangeint modernizer look for i usage
in the first ancestor of a for loop that is not a label.

Fixes golang/go#76470

Change-Id: I79395efba26ad689514b7841d343932448ea9318
Reviewed-on: https://go-review.googlesource.com/c/tools/+/724960
Auto-Submit: Nicholas Husin <husin@google.com>
Reviewed-by: Nicholas Husin <husin@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Josh Rickmar <jrick@zettaport.com>