go/analysis/passes/modernize: fix stringscut false positives
This CL narrows the set of binary expressions (involving
the index "i" returned by strings.Index) for which we
suggest a stringscut modernization.

Before, we determined if the binary expression establishes
that index "i" is negative or non-negative.
Now, we check that the condition is equivalent to i < 0
or i >= 0 by matching to an exact set of expressions.

Suggesting a modernization when i is used in
expressions like i > 1, or i > 0, may result in an invalid fix
if the code intends to assert something about the location of the
substring rather than just its existence.

Fixes golang/go#76687

Change-Id: I598278a2d6751b1c2b2973099809c68d91832eb7
Reviewed-on: https://go-review.googlesource.com/c/tools/+/726840
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Madeline Kalil <mkalil@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>