go/analysis/passes/modernize: fix stringsbuilder bug with in var(...)
The previous logic neglected to consider the possibility of a var decl
with explicit parens, and inserted a statement where it wouldn't go.
Now, we check for parens, and if so, explicitly move the close paren
(by insertion and deletion).

Byte-level surgery is good for avoiding comment loss, but it is so
fiddly and low level.  We really need a better calculus of edits.

Fixes golang/go#76934

Change-Id: Ib21c9cd379035e0fdb9c098e7682e38347bdf81c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/731881
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Peter Weinberger <pjw@google.com>